\name{dcor.ttest}
\alias{dcor.ttest}
\alias{dcor.t}
\title{ Distance Correlation t-test for High Dimensions}
\description{Defunct: use \code{dcorT.test} and \code{dcorT}.}
\usage{
dcor.t(x, y, distance = FALSE)
dcor.ttest(x, y, distance = FALSE)
}
\arguments{
  \item{x}{ data or distances of first sample}
  \item{y}{ data or distances of second sample}
  \item{distance}{ TRUE if x and y are distances, otherwise FALSE}
}
\details{
See \code{\link{dcorT}}. 
}
\author{ Maria L. Rizzo \email{mrizzo@bgsu.edu} and
Gabor J. Szekely
}
\keyword{ htest }
\keyword{ multivariate }
\concept{ energy statistics }

