\name{energy-package}
\alias{energy-package}
\alias{energy}
\docType{package}
\title{
 E-statistics: Multivariate Inference via the Energy of Data
}
\description{
Description: E-statistics (energy) tests and statistics for multivariate and univariate inference,
             including distance correlation, one-sample, two-sample, and multi-sample tests for
             comparing multivariate distributions, are implemented. Measuring and testing
             multivariate independence based on distance correlation, partial distance
             correlation, multivariate goodness-of-fit tests, clustering based on energy distance,
             testing for multivariate normality, distance components (disco) for non-parametric
             analysis of structured data, and other energy statistics/methods are implemented.
}
\author{
Maria L. Rizzo and Gabor J. Szekely
}
\references{
G. J. Szekely and M. L. Rizzo (2013). Energy statistics:
A class of statistics based on distances, \emph{Journal of
Statistical Planning and Inference}.

M. L. Rizzo and G. J. Szekely (2016). Energy Distance,
\emph{WIRES Computational Statistics}, Wiley, Volume 8 Issue 1, 27-38.
Available online Dec., 2015, \doi{10.1002/wics.1375}.

G. J. Szekely and M. L. Rizzo (2017). The Energy of Data. 
\emph{The Annual Review of Statistics and Its Application} 
4:447-79. 

G. J. Szekely and M. L. Rizzo (2023). \emph{The Energy of Data and Distance Correlation}. Chapman & Hall/CRC Monographs on Statistics and Applied Probability. ISBN 9781482242744. 
\url{https://www.routledge.com/The-Energy-of-Data-and-Distance-Correlation/Szekely-Rizzo/p/book/9781482242744}.
}
\keyword{ package }
\keyword{ multivariate }
