\name{energy-deprecated}
\alias{DCOR}
\title{ Deprecated Functions}
\description{ These deprecated functions have been replaced by revised functions and will be removed in future releases of the energy package.}
\usage{
DCOR(x, y, index=1.0)
}
\arguments{
  \item{x}{ data or distances of first sample}
  \item{y}{ data or distances of second sample}
  \item{index}{ exponent on Euclidean distance in (0, 2)}
}
\details{ 
DCOR is an R version replaced by faster compiled code.
}
\keyword{ multivariate }
\keyword{ nonparametric }
\concept{ independence }
\concept{ multivariate }
\concept{ distance correlation }
\concept{ distance covariance }
\concept{ energy statistics }

