\name{energy-package}
\alias{energy-package}
\alias{energy}
\docType{package}
\title{
 E-statistics (energy statistics)
}
\description{
Description: E-statistics (energy) tests and statistics for comparing distributions: multivariate normality, multivariate distance components and k-sample test for equal distributions, hierarchical clustering by e-distances, multivariate independence tests, distance correlation, goodness-of-fit tests. Energy-statistics concept based on a generalization of Newton's potential energy is due to Gabor J. Szekely.
}
\author{
Maria L. Rizzo and Gabor J. Szekely
}
\references{
G. J. Szekely and M. L. Rizzo (2013). Energy statistics:
A class of statistics based on distances, \emph{Journal of
Statistical Planning and Inference},
\url{http://dx.doi.org/10.1016/j.jspi.2013.03.018}
}
\keyword{ package }
\keyword{ multivariate }
